UI viewmodel

+++ app/src/main/java/com/indie/apps/pannypal/repository/UserRepositoryImpl.kt
@@ -8,7 +8,8 @@ class UserRepositoryImpl @Inject constructor(private val userDao: UserDao) : Use
 
     override suspend fun getUser() = userDao.getUser()
 
-    override suspend fun updateAmount(incomeAmt: Long, expenseAmt: Long) = userDao.updateAmount(incomeAmt, expenseAmt)
+    override suspend fun updateAmount(incomeAmt: Double, expenseAmt: Double) =
+        userDao.updateAmount(incomeAmt, expenseAmt)
 
     override suspend fun insert(user: User) = userDao.insert(user)
